refactor(appserver): extract parsing helpers
